# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.RuleCondition:
    properties:
      field:
        description: 匹配字段 (provider/region/tag.xxx/name/attributes.xxx)
        type: string
      operator:
        description: 操作符 (eq/ne/contains/regex/in)
        type: string
      value:
        description: 匹配值
        type: string
    type: object
...
  ginx.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  web.BatchBindReq:
    properties:
      env_id:
        type: integer
      resource_ids:
        items:
          type: integer
        type: array
      resource_type:
        type: string
    required:
    - env_id
    - resource_ids
    - resource_type
    type: object
  web.BindResourceReq:
    properties:
      env_id:
        description: 环境ID
        type: integer
      resource_id:
        type: integer
      resource_type:
        description: instance/asset
        type: string
    required:
    - env_id
    - resource_id
    - resource_type
    type: object
  web.BindingVO:
    properties:
      bind_type:
        type: string
      create_time:
        type: integer
      env_id:
        type: integer
      id:
        type: integer
      node_id:
        type: integer
      resource_id:
        type: integer
      resource_type:
        type: string
      rule_id:
        type: integer
    type: object
...
  web.CreateNodeReq:
    properties:
      description:
        description: 描述
        type: string
      name:
        description: 节点名称
        type: string
      order:
        description: 排序
        type: integer
      owner:
        description: 负责人
        type: string
      parent_id:
        description: 父节点ID
        type: integer
      tags:
        description: 标签
        items:
          type: string
        type: array
      team:
        description: 团队
        type: string
      uid:
        description: 节点唯一标识
        type: string
    required:
    - name
    type: object
  web.CreateRuleReq:
    properties:
      conditions:
        items:
          $ref: '#/definitions/domain.RuleCondition'
        type: array
      description:
        type: string
      enabled:
        type: boolean
      env_id:
        type: integer
      name:
        type: string
      node_id:
        type: integer
      priority:
        type: integer
    required:
    - conditions
    - env_id
    - name
    - node_id
    type: object
...
  web.MoveNodeReq:
    properties:
      new_parent_id:
        type: integer
    type: object
  web.NodeVO:
    properties:
      create_time:
        type: integer
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      order:
        type: integer
      owner:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      team:
        type: string
      uid:
        type: string
      update_time:
        type: integer
    type: object
...
  web.RuleVO:
    properties:
      conditions:
        items:
          $ref: '#/definitions/domain.RuleCondition'
        type: array
      create_time:
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      env_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      node_id:
        type: integer
      priority:
        type: integer
      update_time:
        type: integer
    type: object
...
  web.TreeNodeVO:
    properties:
      children:
        items:
          $ref: '#/definitions/web.TreeNodeVO'
        type: array
      create_time:
        type: integer
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      order:
        type: integer
      owner:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      resource_count:
        type: integer
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      team:
        type: string
      uid:
        type: string
      update_time:
        type: integer
    type: object
...
  web.UpdateNodeReq:
    properties:
      description:
        type: string
      name:
        type: string
      order:
        type: integer
      owner:
        type: string
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      team:
        type: string
      uid:
        type: string
    required:
    - name
    type: object
...
  web.UpdateRuleReq:
    properties:
      conditions:
        items:
          $ref: '#/definitions/domain.RuleCondition'
        type: array
      description:
        type: string
      enabled:
        type: boolean
      env_id:
        type: integer
      name:
        type: string
      node_id:
        type: integer
      priority:
        type: integer
    required:
    - conditions
    - env_id
    - name
    - node_id
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业云资产管理服务 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: E-CAM Service API
  version: "1.0"
paths:
...
  /api/v1/cam/service-tree/bindings/{id}:
    delete:
      parameters:
      - description: 绑定ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 解绑资源
      tags:
      - 服务树
...
  /api/v1/cam/service-tree/nodes:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 父节点ID
        in: query
        name: parent_id
        type: integer
      - description: 层级
        in: query
        name: level
        type: integer
      - description: 名称
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.NodeVO'
                  type: array
              type: object
      summary: 获取节点列表
      tags:
      - 服务树
    post:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 节点信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateNodeReq'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      summary: 创建服务树节点
      tags:
      - 服务树
  /api/v1/cam/service-tree/nodes/{id}:
    delete:
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 删除节点
      tags:
      - 服务树
    get:
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  $ref: '#/definitions/web.NodeVO'
              type: object
      summary: 获取节点详情
      tags:
      - 服务树
    put:
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.UpdateNodeReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 更新节点
      tags:
      - 服务树
  /api/v1/cam/service-tree/nodes/{id}/bindings:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 环境ID
        in: query
        name: env_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.BindingVO'
                  type: array
              type: object
      summary: 获取节点绑定的资源
      tags:
      - 服务树
    post:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 资源信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.BindResourceReq'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      summary: 绑定资源到节点
      tags:
      - 服务树
  /api/v1/cam/service-tree/nodes/{id}/bindings/batch:
    post:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 资源列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.BatchBindReq'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      summary: 批量绑定资源
      tags:
      - 服务树
  /api/v1/cam/service-tree/nodes/{id}/move:
    put:
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 目标父节点
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.MoveNodeReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 移动节点
      tags:
      - 服务树
  /api/v1/cam/service-tree/resources/{type}/{id}/node:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 资源类型
        in: path
        name: type
        required: true
        type: string
      - description: 资源ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  $ref: '#/definitions/web.NodeVO'
              type: object
      summary: 获取资源所属节点
      tags:
      - 服务树
  /api/v1/cam/service-tree/rules:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 节点ID
        in: query
        name: node_id
        type: integer
      - description: 是否启用
        in: query
        name: enabled
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.RuleVO'
                  type: array
              type: object
      summary: 获取规则列表
      tags:
      - 服务树
    post:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateRuleReq'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      summary: 创建绑定规则
      tags:
      - 服务树
  /api/v1/cam/service-tree/rules/{id}:
    delete:
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 删除规则
      tags:
      - 服务树
    get:
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  $ref: '#/definitions/web.RuleVO'
              type: object
      summary: 获取规则详情
      tags:
      - 服务树
    put:
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      - description: 规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.UpdateRuleReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 更新规则
      tags:
      - 服务树
  /api/v1/cam/service-tree/rules/execute:
    post:
      description: 手动触发规则引擎，对未绑定资源执行自动匹配，绑定到规则指定的环境
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      summary: 执行规则匹配
      tags:
      - 服务树
  /api/v1/cam/service-tree/tree:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 根节点ID
        in: query
        name: root_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  $ref: '#/definitions/web.TreeNodeVO'
              type: object
      summary: 获取服务树结构
      tags:
      - 服务树
...
schemes:
- http
- https
swagger: "2.0"